libraries

+++ common/include/common.h
@@ -88,7 +88,7 @@ inline PPMImage *readPPM(const char *filename)
     }
 
     //read pixel data from file
-    if (fread(img->data, 3 * img->x, img->y, fp) != img->y) {
+    if ((int) fread(img->data, 3 * img->x, img->y, fp) != img->y) {
          fprintf(stderr, "Error loading image '%s'\n", filename);
          exit(1);
     }
